pyproject.toml
[project]
name = "openai-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Kunal7069",email = "jaink7069@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "openai (>=1.88.0,<2.0.0)",
    "tokencost (>=0.1.23,<0.2.0)"
]

[tool.poetry]
packages = [{include = "openai_tracker", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
